(Improve) Recoil optimization
