+++ src/components/App.js
@@ -20,6 +20,7 @@ import Notice from './myPage/Notice';
 import Search from '../pages/Search';
 
 function App() {
+  const test = useRecoilValue(authState);
   const { isFirstLogin, isLogin } = useRecoilValue(authState);
   const location = useLocation();
   const history = useHistory();
@@ -44,7 +45,7 @@ function App() {
       history.replace('/monster');
       return;
     }
-  }, [isFirstLogin, isLogin, location.pathname]);
+  }, []);
 
   return (
     <Layout>
