+++ src/components/habit/TodayHabitList.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useRecoilValue } from 'recoil';
 import styled from 'styled-components';
-import { habitIdListState } from '../../recoil/states/test';
+import { habitIdListState } from '../../recoil/states/habit';
 
 import { TodayHabit, NoHabitHelper } from './';
