+++ src/components/login/GoogleLogin.js
@@ -9,6 +9,7 @@ import { GoogleSymbol } from '../../assets/icons/loginSymbol';
 import { OK } from '../../constants/statusCode';
 
 const GoogleLogin = () => {
+  console.log('googleLoginRender');
   const history = useHistory();
   const googleLoginBtn = useRef(null);
   const socialName = 'google';
